Remove reliance on bundled GPU plugin
This reworks everything to instead intercept whichever DrawCallbacks are
currently set by another plugin, such as GPU or 117 HD. Draw calls are
passed on to the original DrawCallback handler only when the blindfold
plugin wants a thing to be drawn. Finally, to hide the UI, an overlay is
used to clear the UI to be fully transparent.

What this all means is that the plugin will now support any GPU plugin,
as long as it uses DrawCallbacks in the usual fashion, and the plugin
won't need to be kept up to date with the GPU plugin's more frequent
changes.